lifetime_sprite_system and dead_sprite_system removed
both could create a race condition, with lifetime_system and
dead_system, which resulted in sprites not being killed from the sprite
group.

Since this cannot be resolved within the library, and the names
suggested a safety that doesn't exist, both have been removed.

Instead, remove_component now checks if the component has a 'shutdown'
attribute.  'shutdown' is expected to be a zero-parameter functions that
will be executed if it exists.