Updated ESprite, added EVsprite
ESprite now creates an empty default image and a rect placing it 1px
outside the screen, so that every required attribute is available at
instantiation time.

EVsprite is a "virtual" sprite where `image` is a property provided by
an `image_factory` passed to `init`.

The interface to the `image_factory` also is its `image` attribute.